refactor TestUtils@assertSocialUserEquals
