+++ Backend/src/test/java/com/social/seed/repository/SocialUserInterestInHashTagRepositoryTest.java
@@ -17,6 +17,7 @@ package com.social.seed.repository;
 
 import com.social.seed.model.HashTag;
 import com.social.seed.model.SocialUser;
+import com.social.seed.utils.TestUtils;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
